+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 
 # Unreleased
 
-- Add support for AIP-80 compliant private keys.
+- [`Breaking`] Add support for AIP-80 compliant private keys. The `toString` of `Ed25519PrivateKey` and `Secp256k1PrivateKey` will now contain a prefix (e.g. `ed25510-priv-0x123...`, `secp256k1-priv-0x123...`).You can use `toHexString()` to get the Hex string representation instead. 
 
 # 1.30.0 (2024-10-21)